requirements.txt
pyserial==3.4
pandas==1.0.4
numpy==1.15.4
sklearn==0.20.1